Extract `EventsTrait` and `BehaviorsTrait`.
