+++ framework/base/BehaviorsInterface.php
@@ -13,8 +13,74 @@ namespace yii\base;
  * @author Robert Korulczyk <robert@korulczyk.pl>
  * @since 2.0.11
  */
-interface ComponentInterface extends VirtualPropertiesInterface
+interface BehaviorsInterface extends EventsInterface
 {
+
+    /**
+     * Returns a value indicating whether a property is defined for this component.
+     * A property is defined if:
+     *
+     * - the class has a getter or setter method associated with the specified name
+     *   (in this case, property name is case-insensitive);
+     * - the class has a member variable with the specified name (when `$checkVars` is true);
+     * - an attached behavior has a property of the given name (when `$checkBehaviors` is true).
+     *
+     * @param string $name the property name
+     * @param boolean $checkVars whether to treat member variables as properties
+     * @param boolean $checkBehaviors whether to treat behaviors' properties as properties of this component
+     * @return boolean whether the property is defined
+     * @see canGetProperty()
+     * @see canSetProperty()
+     */
+    public function hasProperty($name, $checkVars = true, $checkBehaviors = true);
+
+    /**
+     * Returns a value indicating whether a property can be read.
+     * A property can be read if:
+     *
+     * - the class has a getter method associated with the specified name
+     *   (in this case, property name is case-insensitive);
+     * - the class has a member variable with the specified name (when `$checkVars` is true);
+     * - an attached behavior has a readable property of the given name (when `$checkBehaviors` is true).
+     *
+     * @param string $name the property name
+     * @param boolean $checkVars whether to treat member variables as properties
+     * @param boolean $checkBehaviors whether to treat behaviors' properties as properties of this component
+     * @return boolean whether the property can be read
+     * @see canSetProperty()
+     */
+    public function canGetProperty($name, $checkVars = true, $checkBehaviors = true);
+
+    /**
+     * Returns a value indicating whether a property can be set.
+     * A property can be written if:
+     *
+     * - the class has a setter method associated with the specified name
+     *   (in this case, property name is case-insensitive);
+     * - the class has a member variable with the specified name (when `$checkVars` is true);
+     * - an attached behavior has a writable property of the given name (when `$checkBehaviors` is true).
+     *
+     * @param string $name the property name
+     * @param boolean $checkVars whether to treat member variables as properties
+     * @param boolean $checkBehaviors whether to treat behaviors' properties as properties of this component
+     * @return boolean whether the property can be written
+     * @see canGetProperty()
+     */
+    public function canSetProperty($name, $checkVars = true, $checkBehaviors = true);
+
+    /**
+     * Returns a value indicating whether a method is defined.
+     * A method is defined if:
+     *
+     * - the class has a method with the specified name
+     * - an attached behavior has a method with the given name (when `$checkBehaviors` is true).
+     *
+     * @param string $name the property name
+     * @param boolean $checkBehaviors whether to treat behaviors' methods as methods of this component
+     * @return boolean whether the property is defined
+     */
+    public function hasMethod($name, $checkBehaviors = true);
+
     /**
      * Returns a list of behaviors that this component should behave as.
      *
@@ -41,65 +107,6 @@ interface ComponentInterface extends VirtualPropertiesInterface
      * @return array the behavior configurations.
      */
     public function behaviors();
-
-    /**
-     * Returns a value indicating whether there is any handler attached to the named event.
-     * @param string $name the event name
-     * @return boolean whether there is any handler attached to the event.
-     */
-    public function hasEventHandlers($name);
-
-    /**
-     * Attaches an event handler to an event.
-     *
-     * The event handler must be a valid PHP callback. The following are
-     * some examples:
-     *
-     * ```
-     * function ($event) { ... }         // anonymous function
-     * [$object, 'handleClick']          // $object->handleClick()
-     * ['Page', 'handleClick']           // Page::handleClick()
-     * 'handleClick'                     // global function handleClick()
-     * ```
-     *
-     * The event handler must be defined with the following signature,
-     *
-     * ```
-     * function ($event)
-     * ```
-     *
-     * where `$event` is an [[Event]] object which includes parameters associated with the event.
-     *
-     * @param string $name the event name
-     * @param callable $handler the event handler
-     * @param mixed $data the data to be passed to the event handler when the event is triggered.
-     * When the event handler is invoked, this data can be accessed via [[Event::data]].
-     * @param boolean $append whether to append new event handler to the end of the existing
-     * handler list. If false, the new handler will be inserted at the beginning of the existing
-     * handler list.
-     * @see off()
-     */
-    public function on($name, $handler, $data = null, $append = true);
-
-    /**
-     * Detaches an existing event handler from this component.
-     * This method is the opposite of [[on()]].
-     * @param string $name event name
-     * @param callable $handler the event handler to be removed.
-     * If it is null, all handlers attached to the named event will be removed.
-     * @return boolean if a handler is found and detached
-     * @see on()
-     */
-    public function off($name, $handler = null);
-
-    /**
-     * Triggers an event.
-     * This method represents the happening of an event. It invokes
-     * all attached handlers for the event including class-level handlers.
-     * @param string $name the event name
-     * @param Event $event the event parameter. If not set, a default [[Event]] object will be created.
-     */
-    public function trigger($name, Event $event = null);
 
     /**
      * Returns the named behavior object.
