+++ framework/base/EventsTrait.php
@@ -1,0 +1,193 @@
+<?php
+
+/**
+ * @link http://www.yiiframework.com/
+ * @copyright Copyright (c) 2008 Yii Software LLC
+ * @license http://www.yiiframework.com/license/
+ */
+
+namespace yii\base;
+
+/**
+ *
+ * @author Robert Korulczyk <robert@korulczyk.pl>
+ */
+trait EventsTrait
+{
+
+    use VirtualPropertiesTrait;
+
+    /**
+     * @var array the attached event handlers (event name => handlers)
+     */
+    private $_events = [];
+
+    public function prepareEvents()
+    {
+
+    }
+
+    /**
+     * Sets the value of a component property.
+     * This method will check in the following order and act accordingly:
+     *
+     *  - a property defined by a setter: set the property value
+     *  - an event in the format of "on xyz": attach the handler to the event "xyz"
+     *  - a behavior in the format of "as xyz": attach the behavior named as "xyz"
+     *  - a property of a behavior: set the behavior property value
+     *
+     * Do not call this method directly as it is a PHP magic method that
+     * will be implicitly called when executing `$component->property = $value;`.
+     * @param string $name the property name or the event name
+     * @param mixed $value the property value
+     * @throws UnknownPropertyException if the property is not defined
+     * @throws InvalidCallException if the property is read-only.
+     * @see __get()
+     */
+    public function __set($name, $value)
+    {
+        $setter = 'set' . $name;
+        if (method_exists($this, $setter)) {
+            // set property
+            $this->$setter($value);
+
+            return;
+        } elseif (strncmp($name, 'on ', 3) === 0) {
+            // on event: attach event handler
+            $this->on(trim(substr($name, 3)), $value);
+
+            return;
+        }
+        if (method_exists($this, 'get' . $name)) {
+            throw new InvalidCallException('Setting read-only property: ' . get_class($this) . '::' . $name);
+        } else {
+            throw new UnknownPropertyException('Setting unknown property: ' . get_class($this) . '::' . $name);
+        }
+    }
+
+    /**
+     * This method is called after the object is created by cloning an existing one.
+     * It removes all behaviors because they are attached to the old object.
+     */
+    public function __clone()
+    {
+        $this->_events = [];
+    }
+
+    /**
+     * Returns a value indicating whether there is any handler attached to the named event.
+     * @param string $name the event name
+     * @return boolean whether there is any handler attached to the event.
+     */
+    public function hasEventHandlers($name)
+    {
+        $this->prepareEvents();
+        return !empty($this->_events[$name]) || Event::hasHandlers($this, $name);
+    }
+
+    /**
+     * Attaches an event handler to an event.
+     *
+     * The event handler must be a valid PHP callback. The following are
+     * some examples:
+     *
+     * ```
+     * function ($event) { ... }         // anonymous function
+     * [$object, 'handleClick']          // $object->handleClick()
+     * ['Page', 'handleClick']           // Page::handleClick()
+     * 'handleClick'                     // global function handleClick()
+     * ```
+     *
+     * The event handler must be defined with the following signature,
+     *
+     * ```
+     * function ($event)
+     * ```
+     *
+     * where `$event` is an [[Event]] object which includes parameters associated with the event.
+     *
+     * @param string $name the event name
+     * @param callable $handler the event handler
+     * @param mixed $data the data to be passed to the event handler when the event is triggered.
+     * When the event handler is invoked, this data can be accessed via [[Event::data]].
+     * @param boolean $append whether to append new event handler to the end of the existing
+     * handler list. If false, the new handler will be inserted at the beginning of the existing
+     * handler list.
+     * @see off()
+     */
+    public function on($name, $handler, $data = null, $append = true)
+    {
+        $this->prepareEvents();
+        if ($append || empty($this->_events[$name])) {
+            $this->_events[$name][] = [$handler, $data];
+        } else {
+            array_unshift($this->_events[$name], [$handler, $data]);
+        }
+    }
+
+    /**
+     * Detaches an existing event handler from this component.
+     * This method is the opposite of [[on()]].
+     * @param string $name event name
+     * @param callable $handler the event handler to be removed.
+     * If it is null, all handlers attached to the named event will be removed.
+     * @return boolean if a handler is found and detached
+     * @see on()
+     */
+    public function off($name, $handler = null)
+    {
+        $this->prepareEvents();
+        if (empty($this->_events[$name])) {
+            return false;
+        }
+        if ($handler === null) {
+            unset($this->_events[$name]);
+            return true;
+        } else {
+            $removed = false;
+            foreach ($this->_events[$name] as $i => $event) {
+                if ($event[0] === $handler) {
+                    unset($this->_events[$name][$i]);
+                    $removed = true;
+                }
+            }
+            if ($removed) {
+                $this->_events[$name] = array_values($this->_events[$name]);
+            }
+            return $removed;
+        }
+    }
+
+    /**
+     * Triggers an event.
+     * This method represents the happening of an event. It invokes
+     * all attached handlers for the event including class-level handlers.
+     * @param string $name the event name
+     * @param Event $event the event parameter. If not set, a default [[Event]] object will be created.
+     */
+    public function trigger($name, Event $event = null)
+    {
+        $this->prepareEvents();
+        if (!empty($this->_events[$name])) {
+            if ($event === null) {
+                $event = new Event;
+            }
+            if ($event->sender === null) {
+                $event->sender = $this;
+            }
+            $event->handled = false;
+            $event->name = $name;
+            foreach ($this->_events[$name] as $handler) {
+                $event->data = $handler[1];
+                call_user_func($handler[0], $event);
+                // stop further handling if the event is handled
+                if ($event->handled) {
+                    return;
+                }
+            }
+        }
+        // invoke class-level attached handlers
+        Event::trigger($this, $name, $event);
+    }
+
+}
